refactor: rename sp file
